+++ tests/Feature/Auth/VerificationNotificationTest.php
@@ -41,7 +41,7 @@ class VerificationNotificationTest extends TestCase
 
         $this->actingAs($user)
             ->post(route('verification.send'))
-            ->assertRedirect(route('dashboard', absolute: false));
+            ->assertRedirect(route('evolayer.base.home', absolute: false));
 
         Notification::assertNothingSent();
     }